[ADD] petition status info inquiry api

+++ src/main/kotlin/io/github/monsteel/petition/util/Constant.kt
@@ -3,5 +3,5 @@ package io.github.monsteel.petition.util
 object Constant {
     const val MILLISECONDS_FOR_A_HOUR: Int = 1000 * 60 * 60
     const val PETITION_VALIDITY_DAY:Int = 30
-    const val DO_NOT_MODIFY_AGREE_COUNT:Int = 30
+    const val DO_NOT_MODIFY_AGREE_COUNT:Long = 30
 }
